+++ src/app/cmd_result.rs
@@ -17,6 +17,16 @@ pub enum HDir {
     Right,
 }
 
+/// the symbolic reference to the panel to close
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum PanelReference {
+    Active,
+    Leftest,
+    Rightest,
+    Id(PanelId),
+    Preview,
+}
+
 /// Result of applying a command to a state
 pub enum AppStateCmdResult {
     ApplyOnPanel {
@@ -24,7 +34,7 @@ pub enum AppStateCmdResult {
     },
     ClosePanel {
         validate_purpose: bool,
-        id: Option<PanelId>, // None if current panel
+        panel_ref: PanelReference,
     },
     DisplayError(String),
     ExecuteSequence {