+++ src/lib.rs
@@ -9,7 +9,6 @@
 
 pub mod app;
 pub mod browser;
-pub mod clap;
 pub mod cli;
 pub mod command;
 pub mod conf;